requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
amqp==2.6.1
    # via kombu
aniso8601==7.0.0
    # via graphene
asgiref==3.5.0
    # via django
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   snowflake-connector-python
async-timeout==4.0.2
    # via redis
bcrypt==3.2.0
    # via paramiko
billiard==3.6.4.0
    # via celery
bleach==5.0.0
    # via -r requirements.in
boto3==1.21.41
    # via -r requirements.in
botocore==1.24.41
    # via
    #   boto3
    #   s3transfer
cachetools==5.0.0
    # via google-auth
celery[redis]==4.4.6
    # via -r requirements.in
certifi==2021.10.8
    # via
    #   elasticsearch
    #   requests
    #   snowflake-connector-python
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
    #   snowflake-connector-python
charset-normalizer==2.0.12
    # via
    #   requests
    #   snowflake-connector-python
cryptography==36.0.2
    # via
    #   -r requirements.in
    #   paramiko
    #   pyopenssl
    #   snowflake-connector-python
cx-oracle==8.3.0
    # via -r requirements.in
cython==0.29.21
    # via -r requirements.in
deprecated==1.2.13
    # via redis
django==3.2.12
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-dirtyfields
    #   django-filter
    #   django-graphql-jwt
    #   django-guardian
    #   django-storages
    #   djangorestframework
    #   graphene-django
django-cors-headers==3.11.0
    # via -r requirements.in
django-dirtyfields==1.8.1
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-graphql-jwt==0.3.1
    # via -r requirements.in
django-guardian==2.4.0
    # via -r requirements.in
django-ordered-model==3.5
    # via -r requirements.in
django-storages==1.12.3
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
djangorestframework-jwt==1.11.0
    # via -r requirements.in
dnspython==2.2.1
    # via -r requirements.in
elasticsearch==7.17.2
    # via
    #   -r requirements.in
    #   elasticsearch-dsl
elasticsearch-dsl==7.4.0
    # via -r requirements.in
future==0.18.2
    # via celery
google-api-core[grpc]==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
google-auth==2.6.4
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-bigquery==3.0.1
    # via -r requirements.in
google-cloud-bigquery-storage==2.13.1
    # via google-cloud-bigquery
google-cloud-core==2.3.0
    # via google-cloud-bigquery
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.2
    # via google-cloud-bigquery
googleapis-common-protos==1.56.0
    # via
    #   google-api-core
    #   grpcio-status
graphene==2.1.9
    # via graphene-django
graphene-django==2.15.0
    # via
    #   -r requirements.in
    #   django-graphql-jwt
graphql-core==2.3.2
    # via
    #   django-graphql-jwt
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==2.0.1
    # via graphene
grpcio==1.44.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.44.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via
    #   requests
    #   snowflake-connector-python
importlib-metadata==4.11.3
    # via
    #   kombu
    #   redis
isodate==0.6.1
    # via python3-saml
jinja2==3.1.1
    # via -r requirements.in
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
kombu==4.6.11
    # via celery
lxml==4.6.5
    # via
    #   python3-saml
    #   xmlsec
markupsafe==2.1.1
    # via jinja2
mysql-connector-python==8.0.28
    # via -r requirements.in
numpy==1.21.6
    # via
    #   -r requirements.in
    #   pandas
    #   pyarrow
oscrypto==1.3.0
    # via snowflake-connector-python
packaging==21.3
    # via
    #   google-cloud-bigquery
    #   redis
pandas==1.3.5
    # via -r requirements.in
paramiko==2.10.3
    # via sshtunnel
promise==2.3
    # via
    #   graphene-django
    #   graphql-core
    #   graphql-relay
proto-plus==1.20.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
protobuf==3.20.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpcio-status
    #   mysql-connector-python
    #   proto-plus
psycopg2-binary==2.9.3
    # via -r requirements.in
pyarrow==6.0.1
    # via google-cloud-bigquery
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pycryptodomex==3.14.1
    # via snowflake-connector-python
pyjwt==1.7.1
    # via
    #   django-graphql-jwt
    #   djangorestframework-jwt
    #   snowflake-connector-python
pymssql==2.2.5
    # via -r requirements.in
pynacl==1.5.0
    # via paramiko
pyopenssl==21.0.0
    # via snowflake-connector-python
pyparsing==3.0.8
    # via packaging
python-dateutil==2.8.2
    # via
    #   botocore
    #   elasticsearch-dsl
    #   google-cloud-bigquery
    #   pandas
python3-saml==1.14.0
    # via -r requirements.in
pytz==2022.1
    # via
    #   celery
    #   django
    #   djangorestframework
    #   pandas
    #   snowflake-connector-python
redis==4.2.2
    # via celery
requests==2.27.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   snowflake-connector-python
rsa==4.8
    # via google-auth
rx==1.6.1
    # via graphql-core
s3transfer==0.5.2
    # via boto3
singledispatch==3.7.0
    # via graphene-django
six==1.16.0
    # via
    #   bcrypt
    #   bleach
    #   elasticsearch-dsl
    #   google-auth
    #   graphene
    #   graphene-django
    #   graphql-core
    #   graphql-relay
    #   grpcio
    #   isodate
    #   paramiko
    #   promise
    #   pyopenssl
    #   python-dateutil
    #   singledispatch
snowflake-connector-python==2.7.3
    # via -r requirements.in
sqlparse==0.4.2
    # via
    #   -r requirements.in
    #   django
sshtunnel==0.4.0
    # via -r requirements.in
text-unidecode==1.3
    # via graphene-django
typing-extensions==4.1.1
    # via
    #   asgiref
    #   async-timeout
    #   importlib-metadata
    #   redis
urllib3==1.26.9
    # via
    #   botocore
    #   elasticsearch
    #   requests
vine==1.3.0
    # via
    #   amqp
    #   celery
webencodings==0.5.1
    # via bleach
wrapt==1.14.0
    # via deprecated
xmlsec==1.3.12
    # via python3-saml
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
